Moved get_best_matching_color_with_paxinos to utils.py
WIP
